==> music_rating_project/2_nlp_local.ipynb <==
import numpy as np
import pandas as pd
import os
import re
import string
import spacy
from spacy.lang.en import English
from spacy.lang.en.stop_words import STOP_WORDS
from nltk.tokenize import RegexpTokenizer
from sklearn.feature_extraction.text import CountVectorizer, TfidfTransformer
from sklearn.model_selection import train_test_split
from sklearn.model_selection import cross_val_score
from sklearn.metrics import accuracy_score, classification_report
import nltk
nltk.download('words')
nlp = spacy.load('en_core_web_sm')
words = set(nltk.corpus.words.words())

from alphabet_detector import AlphabetDetector
ad = AlphabetDetector()

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

# nlp = English()

# grab all files from webscraped json files in lyrics folder

file_list = []

for f in os.listdir('lyrics'):
    file_list.append(f'lyrics/{f}')
    
file_list.sort()

# read json files into dataframes, and concat them together

df_list = []

for filepath in file_list:
    df = pd.read_json(filepath)
    df_list.append(df)
    
lyrics_df = pd.concat(df_list, axis = 0)

lyrics_df.shape

music_df = pd.read_csv('all_music.csv')

music_df.columns

# imputing null skip, play, and rating values with 0s

music_df['Skips'] = music_df['Skips'].fillna(0)
music_df['Plays'] = music_df['Plays'].fillna(0)
music_df['My Rating'] = music_df['My Rating'].fillna(0)

def convert_rating(df, skip_threshold):
    '''
    this function modifies training data by labeling songs that had a rating as a like, and songs that were rated as 0 and had skips above a threshold as a dislike
    '''
    converted_list = []
    for i, row in df.iterrows():
        if (row['My Rating'] > 0):
            converted_list.append(1)
        elif (row['My Rating'] ==  0) & (row['Skips'] >= skip_threshold):
            converted_list.append(0)
        else:
            converted_list.append(None)
            
    return converted_list

lyrics_df['liked'] = convert_rating(music_df, 1)

lyrics_df['liked'].value_counts()

lyrics_df = lyrics_df.reset_index(drop = True)

lyrics_df = lyrics_df.drop(columns = ['order'])

final_df = lyrics_df.copy()

final_df.head()

final_df[final_df['lyrics'].isna()].head()

# dropping any songs from training dataset that have no lyrics

final_df = final_df[final_df['lyrics'].notnull()]

final_df.head()

def clean_string(lyric):
    '''
    this function takes the lyrics string as an argument,
    removes punctuation, special newline/tab strings, removes numbers, removes stopwords, lemmatizes words, and removes non-latin alphabet characters
    '''
    try:
        
        if lyric == None:
            return None
        else:
            lyric = re.sub('\[.*\]|\(|\)', '', lyric) # remove punct
            lyric = re.sub('\s+', ' ', lyric) # get rid of newline stuff
            modified_punct = '!"#$%&\()*+,-./:;<=>?@[\\]^_`{|}~'
            lyric = [char.lower() for char in lyric if char not in modified_punct + string.digits] # remove more punct and digits
            lyric = ''.join(lyric) #rejoin list of characters back into words
            lyric = lyric[1:] # removing first space in all lyrics
            
            tokenized_lyric = nlp(lyric) #nlp toeknization
            
            nlped_words = [] # removing all stopwords/lemmatize
            
            for word in tokenized_lyric: # lematize
                if word.is_stop == False:
                    nlped_words.append(word.lemma_)
            
            lyric = ' '.join(nlped_words) # rejoin words into string
            lyric = ' '.join(w for w in nltk.wordpunct_tokenize(lyric) if w.lower() in words or not w.isalpha()) # get rid of anything that's not a word
            lyric = ''.join(w for w in lyric if ad.is_latin(w)) # get rid of words containing non-latin characters
            
            if lyric[-1] == ' ':
                lyric = lyric[:-1]
                return lyric
            else:
                return lyric
            
    except Exception as e: # for handling errors where cleaning results in an empty string
        if len(lyric) == 0:
            return None
        else:
            return 'error cleaning'

def clean_list(list_of_lyrics):
    '''
    this function takes a list of lyric strings as an argument and returns a cleaned list
    '''
    
    index = range(len(list_of_lyrics))
    
    error_indices = []
    cleaned_list = []
    
    for i, lyric in zip(index, list_of_lyrics):
        if clean_string(lyric) == 'error cleaning':
            error_indices.append(i)   
        else:
            cleaned_list.append(clean_string(lyric))
            if (i+1) % 10000 == 0:
                print(f'{i+1} done...')
            
    print(error_indices)
            
    return cleaned_list

final_df.shape

# liked song ratios

final_df['liked'].value_counts()

final_df['cleaned_lyrics'] = clean_list(final_df['lyrics'])

# liked songs that actually have lyrics after cleaning

final_df = final_df[final_df['cleaned_lyrics'].notnull()]

final_df['liked'].value_counts(dropna = False)

final_df.to_csv('cleaned_lemmatized_unstopped_df.csv', index = False)
