==> music_rating_project/4_model_selection_local.ipynb <==
import numpy as np
import pandas as pd
# import pandas_profiling as pp
import os
import re
import boto3
import string
import pickle
import tarfile
from sklearn.compose import ColumnTransformer
from sklearn.feature_selection import VarianceThreshold
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.naive_bayes import MultinomialNB
from sklearn.linear_model import LogisticRegression
from sklearn.svm import SVC
from sklearn.neighbors import KNeighborsClassifier
from xgboost import XGBClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split, cross_val_score, GridSearchCV, RandomizedSearchCV, StratifiedKFold
from sklearn.metrics import roc_auc_score, classification_report
from sklearn.pipeline import Pipeline

pd.set_option('display.max_rows', None)
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

### download data from s3

s3_location = 's3://djk-ml-sagemaker/music_lyrics/'

final_df = pd.read_csv(f'{s3_location}cleaned_lemmatized_unstopped_df.csv')

final_df.head()

final_df.liked.value_counts(dropna = False)

# final_df.profile_report()

### creating training/testing datasets, only choosing liked/not liked songs, we will apply the finalized model to unrated music to get recommendations

non_null_df = final_df[final_df['liked'].notnull()].copy()

non_null_df.shape

# this will be the pool of songs that will be recommended to us

null_df = final_df[final_df['liked'].isnull()].copy()

X, y = non_null_df['cleaned_lyrics'], non_null_df['liked']

train_value_counts = y.value_counts()

train_value_counts

majority_count = train_value_counts[train_value_counts.index == 0].iloc[0]
target_count = train_value_counts[train_value_counts.index == 1].iloc[0]

majority_ratio = (majority_count)/(majority_count+target_count)

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = .20, random_state = 42)

### saving training/testing to s3

def write_to_s3(filename, bucket, key):
    '''
    this function uploads files from local directory to s3
    '''
    
    with open(filename,'rb') as f: # Read in binary mode
        return boto3.Session().resource('s3').Bucket(bucket).Object(key).upload_fileobj(f)

# defining s3 bucket/key paths

bucket_name = 'djk-ml-sagemaker'

data_folder = 'music_lyrics'

s3_training_location = f's3://{data_folder}'

train = pd.concat([y_train, X_train], axis = 1)
test = pd.concat([y_test, X_test], axis = 1)

train.to_csv('train.csv', index = None)
test.to_csv('test.csv', index = None)

null_df.to_csv('eligible_song_pool.csv', index = None)

write_to_s3('train.csv', bucket_name, 'music_lyrics/train.csv')

write_to_s3('test.csv', bucket_name, 'music_lyrics/test.csv')

write_to_s3('eligible_song_pool.csv', bucket_name, 'music_lyrics/eligible_song_pool.csv')

### building preprocessing pipeline

# preprocessing steps

tfidf_vectorizer = TfidfVectorizer(stop_words = 'english')
variance_filter = VarianceThreshold(.0005)

word_features = ['cleaned_lyrics']

steps = [
#     ('count_vect', count_vect_no_stops),
    ('tfidf_vectorizer', tfidf_vectorizer),
    ('variance_filter', variance_filter) # removes low variance columns from dataset
]

word_transformer = Pipeline(steps)

# gridsearch params/pipeline

word_transformer.fit_transform(X_train)

all_columns = word_transformer.named_steps['tfidf_vectorizer'].get_feature_names()
filtered_columns = word_transformer.named_steps['variance_filter'].get_support(indices = True)

vectorized_X_train = pd.DataFrame(
    word_transformer.fit_transform(X_train).toarray(),
    columns = [all_columns[i] for i in filtered_columns]
)

vectorized_X_train.head()
vectorized_X_train.shape

### quick model selection. Choosing a handful of classificaiton algorithms with default params to shortlist the promising ones

def model_selection(X_train, y_train, list_of_models:list):
    '''
    this functions takes in a list of classification algorithms with default hyperparam settings
    so we can find the ones that are the most promising
    '''
    
    roc_auc_list = []
    
    importances = []
    
    for model in list_of_models:
        model.fit(X_train, y_train)
        
        feature_importance_dict = {}
        
        try:
            for feature, importance in zip(X_train.columns, model.feature_importances_):
                feature_importance_dict[feature] = importance
        except:
            pass
            
        importance_df = pd.DataFrame.from_dict(feature_importance_dict, orient = 'index')
        
        importances.append(importance_df)
        
        quick_roc_auc = cross_val_score(
            model,
            X_train,
            y_train,
            scoring = 'roc_auc',
            cv = StratifiedKFold(10)
        )
        
        roc_auc_list.append(np.mean(quick_roc_auc))
    
    return roc_auc_list, importances

list_of_models = [
    MultinomialNB(),
    LogisticRegression(random_state = 42),
    SVC(random_state = 42),
    RandomForestClassifier(random_state = 42),
    KNeighborsClassifier(),
    XGBClassifier(random_state = 42)
]

model_selection_roc_aucs, importances = model_selection(vectorized_X_train, y_train, list_of_models)

### evaluating promising models

model_selection_roc_aucs

pd.set_option('display.max_rows', None)

random_forest_top_features = importances[3][0].sort_values(ascending = False)[:50]
xgb_top_features = importances[5][0].sort_values(ascending = False)[:50]

### random forest and xgboost are coming out ahead, looking at feature importances

X_train.head()

random_forest_top_features[:20]
xgb_top_features[:20]

top_feature_overlap = set(random_forest_top_features.index) & set(xgb_top_features.index)

top_feature_overlap

len(top_feature_overlap)/50

#### there is a 32% overlap in the top 50 features for random forest and xgboost

vectorized_X_train.to_csv('vectorized_train.csv')

### taking the top 2 performing models (random forest and xgboost) and optimizing hyperparameters (randomizedsearch) for them

def find_best_model(model_hyperparam_dict, X_train, y_train):
    '''
    this function finds the best model.
    takes in a dictionary with multiple algorithms and hyperparameters
    returns the best model based on roc_auc cross val scores
    also leaves a file with best params
    '''
    
    keys = [k for k, v in model_hyperparam_dict.items()]
    
    cv_results = []
    best_scores = []
    best_estimators = []
    classification_reports = []
    
    for key in keys:
        
        model = model_hyperparam_dict[key]['model']
        hyperparams = model_hyperparam_dict[key]['hyperparams']
        
        print(f'randomsearching {key}...')
        print(f'using hyperparams: \n{hyperparams}')
        
        grid = RandomizedSearchCV(
            estimator = model,
            param_distributions = hyperparams,
            n_iter = 100,
            scoring = 'roc_auc',
            cv = StratifiedKFold(10),
            verbose = 1
        )
        
        grid.fit(X_train, y_train)
        
        cv_results.append(grid.cv_results_)
        best_scores.append(grid.best_score_)
        best_estimators.append(grid.best_estimator_)
        
    return cv_results, best_scores, best_estimators

random_forest_hyperparams = {
    'bootstrap':[True, False],
    'max_depth':[10, 20, 30, 40, 50, 60, 70, 80, 90, 100, None],
    'max_features':['auto', 'sqrt'],
    'min_samples_leaf':[1, 2, 4],
    'min_samples_split':[2, 5, 10],
    'n_estimators':[100, 200, 400, 600, 800, 1000, 1200, 1400, 1600, 1800, 2000]
}

xgb_hyperparams = {
    'max_depth':range(1, 15),
    'learning_rate':[0.0001, 0.001, 0.01, 0.1, 0.2, 0.3],
    'n_estimators':[100, 200, 400, 600, 800, 1000, 1200, 1400, 1600, 1800, 2000],
    'gamma':[0, .01, .1, 1],
    'min_child_weight':[1, 3, 5, 7],
    'colsample_bytree':[.3, .4, .5, .6]
}

model_hyperparam_dict = {
    'random_forest' : {
        'model': RandomForestClassifier(random_state = 42),
        'hyperparams' : random_forest_hyperparams
    },
    
    'xgboost' : {
        'model':XGBClassifier(random_state = 42),
        'hyperparams' : xgb_hyperparams
    }
    
}

cv_results, best_scores, best_estimators = find_best_model(model_hyperparam_dict, vectorized_X_train, y_train)

best_scores
best_estimators

#### appears to perform better than chance, since the training dataset contains about 50% labeled as liked and unliked

### building a pipeline for both using the optimized hyperparams to evaluate on the test set, since our roc_auc crossval score is close

best_random_forest = best_estimators[0]

best_xgb = best_estimators[1]

second_pipeline_steps = [
    ('tfidf_vectorizer', tfidf_vectorizer), # term frequency document infrequency word vectorizer
    ('variance_filter', variance_filter), # removes low variance columns from dataset
    ('classifier', best_random_forest)
]

third_pipeline_steps = [
    ('tfidf_vectorizer', tfidf_vectorizer), # term frequency document infrequency word vectorizer
    ('variance_filter', variance_filter), # removes low variance columns from dataset
    ('classifier', best_xgb)
]

final_random_forest_model = Pipeline(second_pipeline_steps)

final_random_forest_model.fit(X_train, y_train)

final_xgb_model = Pipeline(third_pipeline_steps)
final_xgb_model.fit(X_train, y_train)

random_forest_preds = final_random_forest_model.predict(X_test)
random_forest_probas = final_random_forest_model.predict_proba(X_test)

xgb_preds = final_xgb_model.predict(X_test)
xgb_probas = final_xgb_model.predict_proba(X_test)

### evaluating both models on test set

roc_auc_score(y_test, random_forest_preds)

print(classification_report(y_test, random_forest_preds))

pd.Series([l[1] for l in random_forest_probas]).value_counts(bins = 20).sort_index()

roc_auc_score(y_test, xgb_preds)

print(classification_report(y_test, xgb_preds))

pd.Series([l[1] for l in xgb_probas]).value_counts(bins = 20).sort_index()

### looks like we're losing some % on the eval set for both models, but the xgboost is performing better than randomforest

### saving both models as backups

pickle.dump(final_xgb_model, open('xgb_model.sav', 'wb'))

tar = tarfile.open("xgb_model.tar.gz", "w:gz")
tar.add('xgb_model.sav')

pickle.dump(final_random_forest_model, open('random_forest.sav', 'wb'))

tar = tarfile.open("random_forest.tar.gz", "w:gz")
tar.add('random_forest.sav')
