==> music_rating_project/4c_xgboost_sagemaker_resources.ipynb <==
import boto3
import re
import sagemaker
from sagemaker import get_execution_role

# build inference pipeline
# step 1 - sklearn preprocessing pipeline (tfidf, variance threshold)
# step 2 - xgboost deployment
