==> music_rating_project/5_create_playlist_local.ipynb <==
import numpy as np
import pandas as pd
import pickle
from sklearn.feature_selection import VarianceThreshold
from sklearn.feature_extraction.text import TfidfVectorizer
from xgboost import XGBClassifier
from sklearn.pipeline import Pipeline

pd.set_option('display.max_columns', 500)
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

s3_location = 's3://djk-ml-sagemaker/music_lyrics/'

xgb_model = pickle.load(open('xgb_model.sav', 'rb'))

eligible_songs = pd.read_csv('eligible_song_pool.csv')

eligible_songs.head()

predictions = xgb_model.predict_proba(eligible_songs['cleaned_lyrics'])

preds = xgb_model.predict(eligible_songs['cleaned_lyrics'])

predictions[:5]
preds[:5]

eligible_songs['recommendation'] = [probas[1] for probas in predictions]

eligible_songs.head()

all_music = pd.read_csv(f'{s3_location}all_music.csv')

all_music.head()

playlist_df = pd.merge(all_music, eligible_songs, how = 'inner', left_on = ['Artist', 'Name'], right_on = ['artist_name', 'song_title'])

playlist_df.shape

playlist_df = playlist_df[playlist_df['recommendation'].notnull()]

playlist_df.head()

playlist_df = playlist_df.drop(columns = ['lyrics', 'song_title', 'artist_name', 'liked', 'cleaned_lyrics'])

playlist_df['name_lower'] = [s.lower() for s in playlist_df['Name']]
playlist_df['Artist_lower'] = [s.lower() for s in playlist_df['Artist']]

playlist_df = playlist_df.drop_duplicates(subset = ['name_lower', 'Artist_lower'])

playlist_df.recommendation.value_counts(bins = 20).sort_index(ascending = False)

playlist_df = playlist_df.drop(columns = ['name_lower', 'Artist_lower'])

export_df = playlist_df.sort_values(by = 'recommendation', ascending = False).iloc[:100]

export_df.columns

export_df[['Name', 'Artist', 'recommendation']].head(20)

apple_music_playlist = export_df[['Name', 'Artist', 'Composer', 'Album', 'Grouping', 'Work',
       'Movement Number', 'Movement Count', 'Movement Name', 'Genre', 'Size',
       'Time', 'Disc Number', 'Disc Count', 'Track Number', 'Track Count',
       'Year', 'Date Modified', 'Date Added', 'Bit Rate', 'Sample Rate',
       'Volume Adjustment', 'Kind', 'Equalizer', 'Comments', 'Plays',
       'Last Played', 'Skips', 'Last Skipped', 'My Rating', 'Location']].copy()

with open('recommedations_v0_3','w') as write_tsv:
    write_tsv.write(apple_music_playlist.to_csv(sep='\t', index=False))
