==> music_rating_project/5_upload_to_S3.ipynb <==
import boto3
import json
import os
cred = json.load(open('cred.json'))

access_id = cred['Access key ID']
secret = cred['Secret access key']

s3 = boto3.resource(
    's3',
    aws_access_key_id = access_id,
    aws_secret_access_key = secret
)

# local files

directory = 'data'
files = ['all_music.txt',
         'cleaned_lemmatized_unstopped_df.csv',
         'all_music.csv']

# s3 things
bucket_name = 'djk-ml-sagemaker'

data_folder = 'music_lyrics'

s3_training_location = f's3://{data_folder}'

# training_folder = r'music_lyrics/training/'
# validation_folder = r'music_lyrics/validation/'
# test_folder = r'bikerental/test/'

# s3_model_output_location = r's3://{0}/bikerental/model'.format(bucket_name)
# s3_training_file_location = r's3://{0}/{1}'.format(bucket_name,training_folder)
# s3_validation_file_location = r's3://{0}/{1}'.format(bucket_name,validation_folder)
# s3_test_file_location = r's3://{0}/{1}'.format(bucket_name,test_folder)

print(s3_training_location)

def write_to_s3(filename, bucket, key):
    with open(filename,'rb') as f: # Read in binary mode
        return boto3.Session().resource('s3').Bucket(bucket).Object(key).upload_fileobj(f)

for f in files:
    write_to_s3(f'data/{f}', bucket_name, f'{data_folder}/{f}')
